lift selectBestImgFromWidth to lib
